# docs/swagger.yaml
openapi: 3.0.3
info:
    title: ServeBin
    description:
        Welcome to ServeBin documentation! ServeBin is a cutting-edge HTTP
        testing and debugging tool, built with the latest technologies in Go. This documentation
        provides comprehensive details about the endpoints, parameters, and responses
        offered by ServeBin, empowering developers to streamline their testing workflows
        and ensure the reliability of their applications. Explore the various features
        and capabilities of ServeBin to optimize your development process and elevate
        your HTTP testing experience.
    termsOfService: https://servebin.dev/
    contact:
        email: contact@mrayush.me
    license:
        name: BSD 3-Clause
        url: https://github.com/AyushAgnihotri2025/ServeBin/blob/master/LICENSE
    version: v1.3.0
servers:
    - url: https://servebin.dev/
    - url: https://s1.servebin.dev/
    - url: https://s2.servebin.dev/
...
paths:
    /brotli:
        get:
            tags:
                - Response formats
            summary: Returns Brotli-encoded data.
            description: Returns Brotli-encoded data.
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.BrotliResponse"
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
    /deflate:
        get:
            tags:
                - Response formats
            summary: Returns Deflate-encoded data.
            description: Returns Deflate-encoded data.
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.DeflateResponse"
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
    /deny:
        get:
            tags:
                - Response formats
            summary: Returns page denied by robots.txt rules.
            description: Returns page denied by robots.txt rules.
            responses:
                "200":
                    description: OK
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
...
    /gzip:
        get:
            tags:
                - Response formats
            summary: Returns GZip-encoded data.
            description: Returns GZip-encoded data.
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.GzipResponse"
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                    content: {}
                "500":
                    description: Internal Server Error
                    content:
                        text/plain:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
...
    /html:
        get:
            tags:
                - Response formats
            summary: Returns a simple HTML document.
            description: Returns a simple HTML document.
            responses:
                "200":
                    description: OK
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
...
    /json:
        get:
            tags:
                - Response formats
            summary: Returns a simple JSON document.
            description: Returns a simple JSON document.
            responses:
                "200":
                    description: OK
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                    content: {}
                "500":
                    description: Internal Server Error
                    content:
                        text/plain:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
...
    /robots.txt:
        get:
            tags:
                - Response formats
            summary: Returns some robots.txt rules.
            description: Returns some robots.txt rules.
            responses:
                "200":
                    description: OK
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
    /status/{statuscode}:
        get:
            tags:
                - Status Codes
            summary: Return status code or random status code if more than one are given.
            description: Returns the requester's IP Address.
            parameters:
                - name: statuscode
                  in: path
                  description: Status Code
                  required: true
                  schema:
                      type: integer
                  example: 200
            responses:
                "200":
                    description: OK
                "400":
                    description: Bad Request
                "404":
                    description: Not Found
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
        put:
            tags:
                - Status Codes
            summary: Return status code or random status code if more than one are given.
            description: Returns the requester's IP Address.
            parameters:
                - name: statuscode
                  in: path
                  description: Status Code
                  required: true
                  schema:
                      type: integer
                  example: 200
            responses:
                "200":
                    description: OK
                "400":
                    description: Bad Request
                "404":
                    description: Not Found
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
        post:
            tags:
                - Status Codes
            summary: Return status code or random status code if more than one are given.
            description: Returns the requester's IP Address.
            parameters:
                - name: statuscode
                  in: path
                  description: Status Code
                  required: true
                  schema:
                      type: integer
                  example: 200
            responses:
                "200":
                    description: OK
                "400":
                    description: Bad Request
                "404":
                    description: Not Found
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
        delete:
            tags:
                - Status Codes
            summary: Return status code or random status code if more than one are given.
            description: Returns the requester's IP Address.
            parameters:
                - name: statuscode
                  in: path
                  description: Status Code
                  required: true
                  schema:
                      type: integer
                  example: 200
            responses:
                "200":
                    description: OK
                "400":
                    description: Bad Request
                "404":
                    description: Not Found
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
        patch:
            tags:
                - Status Codes
            summary: Return status code or random status code if more than one are given.
            description: Returns the requester's IP Address.
            parameters:
                - name: statuscode
                  in: path
                  description: Status Code
                  required: true
                  schema:
                      type: integer
                  example: 200
            responses:
                "200":
                    description: OK
                "400":
                    description: Bad Request
                "404":
                    description: Not Found
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
...
    /xml:
        get:
            tags:
                - Response formats
            summary: Returns a simple XML document.
            description: Returns a simple XML document.
            responses:
                "200":
                    description: OK
                "404":
                    description: Not Found
                    content: {}
                "406":
                    description: Not Acceptable
                    content: {}
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
    /zstd:
        get:
            tags:
                - Response formats
            summary: Returns Zstd-compressed data.
            description: Returns Zstd-compressed data.
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.ZstdResponse"
                "404":
                    description: Not Found
                "406":
                    description: Not Acceptable
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
...
components:
    schemas:
...
        response.BrotliResponse:
            type: object
            properties:
                compressed:
                    type: boolean
                    example: true
                headers:
                    type: object
                    additionalProperties:
                        type: string
                ip:
                    type: array
                    items:
                        type: object
        response.DeflateResponse:
            type: object
            properties:
                deflated:
                    type: boolean
                    example: true
                headers:
                    type: object
                    additionalProperties:
                        type: string
                ip:
                    type: array
                    items:
                        type: object
...
        response.GzipResponse:
            type: object
            properties:
                gzipped:
                    type: boolean
                    example: true
                headers:
                    type: object
                    additionalProperties:
                        type: string
                ip:
                    type: array
                    items:
                        type: object
        response.HTTPError:
            type: object
            properties:
                code:
                    type: integer
                    example: 400
                message:
                    type: string
                    example: status bad request
...
        response.ZstdResponse:
            type: object
            properties:
                compressed:
                    type: boolean
                    example: true
                headers:
                    type: object
                    additionalProperties:
                        type: string
                ip:
                    type: array
                    items:
                        type: object
...